+++ firmware/boron-ins-fsm/src/stateMachine.cpp
@@ -19,7 +19,8 @@ unsigned long state1_timer;
 unsigned long state2_duration_timer;
 unsigned long state3_stillness_timer;
 // initialize constants to sensible default values
-unsigned long ins_threshold = INS_THRESHOLD;
+unsigned long stillness_ins_threshold = INS_THRESHOLD;
+unsigned long occupation_detection_ins_threshold = INS_THRESHOLD;
 unsigned long state0_occupant_detection_timer = STATE0_OCCUPANT_DETECTION_TIMER;
 unsigned long state1_max_time = STATE1_MAX_TIME;
 unsigned long state2_max_duration = STATE2_MAX_DURATION;
@@ -64,22 +65,24 @@ void initializeStateMachineConsts() {
     uint16_t initializeConstsFlag;
     uint16_t initializeState3MaxLongStillenssTimeFlag;
     uint16_t initializeState0OccupationDetectionFlag;
+    uint16_t initializeHighConfINSThresholdFlag;
+    uint16_t initializeOccupationDetectionINSThresholdFlag;
 
     // Boron flash memory is initialized to all F's (1's)
     EEPROM.get(ADDR_INITIALIZE_SM_CONSTS_FLAG, initializeConstsFlag);
     Log.info("state machine constants flag is 0x%04X", initializeConstsFlag);
 
-    if (initializeConstsFlag != INITIALIZE_STATE_MACHINE_CONSTS_FLAG) {
-        EEPROM.put(ADDR_INS_THRESHOLD, ins_threshold);
+    if (initializeConstsFlag != INITIALIZATION_FLAG_SET) {
+        EEPROM.put(ADDR_STILLNESS_INS_THRESHOLD, stillness_ins_threshold);
         EEPROM.put(ADDR_STATE1_MAX_TIME, state1_max_time);
         EEPROM.put(ADDR_STATE2_MAX_DURATION, state2_max_duration);
         EEPROM.put(ADDR_STATE3_MAX_STILLNESS_TIME, state3_max_stillness_time);
-        initializeConstsFlag = INITIALIZE_STATE_MACHINE_CONSTS_FLAG;
+        initializeConstsFlag = INITIALIZATION_FLAG_SET;
         EEPROM.put(ADDR_INITIALIZE_SM_CONSTS_FLAG, initializeConstsFlag);
         Log.info("State machine constants were written to flash on bootup.");
     }
     else {
-        EEPROM.get(ADDR_INS_THRESHOLD, ins_threshold);
+        EEPROM.get(ADDR_STILLNESS_INS_THRESHOLD, stillness_ins_threshold);
         EEPROM.get(ADDR_STATE1_MAX_TIME, state1_max_time);
         EEPROM.get(ADDR_STATE2_MAX_DURATION, state2_max_duration);
         EEPROM.get(ADDR_STATE3_MAX_STILLNESS_TIME, state3_max_stillness_time);
@@ -89,14 +92,14 @@ void initializeStateMachineConsts() {
     // Boron flash memory is initialized to all F's (1's)
     // Needs separate intialization for Borons that had versions of the firmware <= 9.3.0
     EEPROM.get(ADDR_INITIALIZE_STATE3_MAX_LONG_STILLNESS_TIME_FLAG, initializeState3MaxLongStillenssTimeFlag);
-    Log.info("state machine constant State3MaxLongStillnessTime flag is 0x%04X", initializeState3MaxLongStillenssTimeFlag);
+    Log.info("State3MaxLongStillnessTime flag is 0x%04X", initializeState3MaxLongStillenssTimeFlag);
 
-    if (initializeState3MaxLongStillenssTimeFlag != INITIALIZE_STATE3_MAX_LONG_STILLNESS_TIME_FLAG) {
+    if (initializeState3MaxLongStillenssTimeFlag != INITIALIZATION_FLAG_SET) {
         // By default, we use the same value for max_stillness_time and max_long_stillness_time
         EEPROM.put(ADDR_STATE3_MAX_LONG_STILLNESS_TIME, state3_max_stillness_time);
         state3_max_long_stillness_time = state3_max_stillness_time;
 
-        initializeState3MaxLongStillenssTimeFlag = INITIALIZE_STATE3_MAX_LONG_STILLNESS_TIME_FLAG;
+        initializeState3MaxLongStillenssTimeFlag = INITIALIZATION_FLAG_SET;
         EEPROM.put(ADDR_INITIALIZE_STATE3_MAX_LONG_STILLNESS_TIME_FLAG, initializeState3MaxLongStillenssTimeFlag);
         Log.info("State machine constant State3MaxLongStillnessTime was written to flash on bootup.");
     }
@@ -107,17 +110,48 @@ void initializeStateMachineConsts() {
 
     // Seperate initialization for State 0 Window
     EEPROM.get(ADDR_INITIALIZE_STATE0_OCCUPANT_DETECTION_TIMER_FLAG, initializeState0OccupationDetectionFlag);
-    Log.info("state machine constant State0OccupationDetectionFlag is 0x%04X", initializeState0OccupationDetectionFlag);
+    Log.info("State0OccupationDetectionFlag is 0x%04X", initializeState0OccupationDetectionFlag);
 
-    if (initializeState0OccupationDetectionFlag != INITIALIZE_STATE0_OCCUPANT_DETECTION_FLAG) {
+    if (initializeState0OccupationDetectionFlag != INITIALIZATION_FLAG_SET) {
         EEPROM.put(ADDR_STATE0_OCCUPANT_DETECTION_TIMER, state0_occupant_detection_timer);
-        initializeState0OccupationDetectionFlag = INITIALIZE_STATE0_OCCUPANT_DETECTION_FLAG;
+        initializeState0OccupationDetectionFlag = INITIALIZATION_FLAG_SET;
         EEPROM.put(ADDR_INITIALIZE_STATE0_OCCUPANT_DETECTION_TIMER_FLAG, initializeState0OccupationDetectionFlag);
         Log.info("State machine constant State0OccupationDetectionTimer was written to flash on bootup.");
     }
     else {
         EEPROM.get(ADDR_STATE0_OCCUPANT_DETECTION_TIMER, state0_occupant_detection_timer);
         Log.info("State machine constant State0OccupationDetectionTimer was read from flash on bootup.");
+    }
+
+    // Initialization for Occupation Detection INS threshold
+    EEPROM.get(ADDR_INITIALIZE_OCCUPANCY_DETECTION_INS_THRESHOLD_FLAG, initializeOccupationDetectionINSThresholdFlag);
+    EEPROM.get(ADDR_INITIALIZE_HIGH_CONF_INS_THRESHOLD_FLAG, initializeHighConfINSThresholdFlag);
+    Log.info("OccupationDetectionINSThresholdFlag is 0x%04X", initializeOccupationDetectionINSThresholdFlag);
+
+    if (initializeOccupationDetectionINSThresholdFlag != INITIALIZATION_FLAG_SET) {
+        // firmware was never v1924
+        if (initializeHighConfINSThresholdFlag != INITIALIZATION_FLAG_HIGH_CONF) {
+            EEPROM.get(ADDR_STILLNESS_INS_THRESHOLD, occupation_detection_ins_threshold);
+            EEPROM.put(ADDR_OCCUPATION_INS_THRESHOLD, occupation_detection_ins_threshold);
+            Log.info("State machine constant OccupantDetectionINSThreshold was written to flash on bootup.");
+        }
+        // firmware was previously v1924
+        else {
+            EEPROM.get(ADDR_STATE3_MAX_LONG_STILLNESS_TIME, state3_max_stillness_time);
+            EEPROM.get(ADDR_STATE3_MAX_STILLNESS_TIME, state3_max_long_stillness_time);
+            EEPROM.get(ADDR_STILLNESS_INS_THRESHOLD, occupation_detection_ins_threshold);
+            EEPROM.get(ADDR_OCCUPATION_INS_THRESHOLD, stillness_ins_threshold);
+            EEPROM.put(ADDR_STILLNESS_INS_THRESHOLD, stillness_ins_threshold);
+            EEPROM.put(ADDR_STATE3_MAX_STILLNESS_TIME, state3_max_stillness_time);
+            EEPROM.put(ADDR_STATE3_MAX_LONG_STILLNESS_TIME, state3_max_long_stillness_time);
+            EEPROM.put(ADDR_OCCUPATION_INS_THRESHOLD, occupation_detection_ins_threshold);
+        }
+        initializeOccupationDetectionINSThresholdFlag = INITIALIZATION_FLAG_SET;
+        EEPROM.put(ADDR_INITIALIZE_OCCUPANCY_DETECTION_INS_THRESHOLD_FLAG, initializeOccupationDetectionINSThresholdFlag);
+    }
+    // firmware is v11000+
+    else {
+        EEPROM.get(ADDR_OCCUPATION_INS_THRESHOLD, occupation_detection_ins_threshold);
     }
 }
 
@@ -145,18 +179,12 @@ void state0_idle() {
     // set the total number of alerts generated to 0
     number_of_alerts_published = 0;
 
-    // do stuff in the state
-    digitalWrite(D2, LOW);
-    digitalWrite(D3, LOW);
-    digitalWrite(D4, LOW);
-    digitalWrite(D5, LOW);
-
     Log.info("You are in state 0, idle: Door status, iAverage = 0x%02X, %f", checkDoor.doorStatus, checkINS.iAverage);
     // default timer to 0 when state doesn't have a timer
     publishDebugMessage(0, checkDoor.doorStatus, checkINS.iAverage, (millis() - timeWhenDoorClosed));
 
     // fix outputs and state exit conditions accordingly
-    if (millis() - timeWhenDoorClosed < state0_occupant_detection_timer && ((unsigned long)checkINS.iAverage > ins_threshold) &&
+    if (millis() - timeWhenDoorClosed < state0_occupant_detection_timer && ((unsigned long)checkINS.iAverage > occupation_detection_ins_threshold) &&
         !isDoorOpen(checkDoor.doorStatus) && !isDoorStatusUnknown(checkDoor.doorStatus)) {
         Log.warn("In state 0, door closed and seeing movement, heading to state 1");
         publishStateTransition(0, 1, checkDoor.doorStatus, checkINS.iAverage);
@@ -184,13 +212,12 @@ void state1_15sCountdown() {
     checkINS = checkINS3331();
 
     // do stuff in the state
-    digitalWrite(D2, HIGH);
     Log.info("You are in state 1, initial countdown: Door status, iAverage, timer = 0x%02X, %f, %ld", checkDoor.doorStatus, checkINS.iAverage,
              (millis() - state1_timer));
     publishDebugMessage(1, checkDoor.doorStatus, checkINS.iAverage, (millis() - state1_timer));
 
     // fix outputs and state exit conditions accordingly
-    if ((unsigned long)checkINS.iAverage > 0 && (unsigned long)checkINS.iAverage < ins_threshold) {
+    if ((unsigned long)checkINS.iAverage > 0 && (unsigned long)checkINS.iAverage < occupation_detection_ins_threshold) {
         Log.warn("no movement, you're going back to state 0 from state 1");
         publishStateTransition(1, 0, checkDoor.doorStatus, checkINS.iAverage);
         saveStateChangeOrAlert(1, 1);
@@ -230,13 +257,12 @@ void state2_duration() {
     char alertMessage[622];
 
     // do stuff in the state
-    digitalWrite(D3, HIGH);
     Log.info("You are in state 2, duration: Door status, iAverage, timer = 0x%02X, %f, %ld", checkDoor.doorStatus, checkINS.iAverage,
              (millis() - state2_duration_timer));
     publishDebugMessage(2, checkDoor.doorStatus, checkINS.iAverage, (millis() - state2_duration_timer));
 
     // fix outputs and state exit conditions accordingly
-    if ((unsigned long)checkINS.iAverage > 0 && (unsigned long)checkINS.iAverage < ins_threshold) {
+    if ((unsigned long)checkINS.iAverage > 0 && (unsigned long)checkINS.iAverage < stillness_ins_threshold) {
         Log.warn("Seeing stillness, going to state3_stillness from state2_duration");
         publishStateTransition(2, 3, checkDoor.doorStatus, checkINS.iAverage);
         saveStateChangeOrAlert(2, 1);
@@ -283,13 +309,12 @@ void state3_stillness() {
     checkINS = checkINS3331();
 
     // do stuff in the state
-    digitalWrite(D4, HIGH);
     Log.info("You are in state 3, stillness: Door status, iAverage, timer = 0x%02X, %f, %ld", checkDoor.doorStatus, checkINS.iAverage,
              (millis() - state3_stillness_timer));
     publishDebugMessage(3, checkDoor.doorStatus, checkINS.iAverage, (millis() - state3_stillness_timer));
 
     // fix outputs and state exit conditions accordingly
-    if ((unsigned long)checkINS.iAverage > ins_threshold) {
+    if ((unsigned long)checkINS.iAverage > stillness_ins_threshold) {
         Log.warn("motion spotted again, going from state3_stillness to state2_duration");
         publishStateTransition(3, 2, checkDoor.doorStatus, checkINS.iAverage);
         saveStateChangeOrAlert(3, 0);
@@ -354,9 +379,9 @@ void publishDebugMessage(int state, unsigned char doorStatus, float INSValue, un
             // from particle docs, max length of publish is 622 chars, I am assuming this includes null char
             char debugMessage[622];
             snprintf(debugMessage, sizeof(debugMessage),
-                     "{\"state\":\"%d\", \"door_status\":\"0x%02X\", \"INS_val\":\"%f\", \"INS_threshold\":\"%lu\", \"timer_status\":\"%lu\", "
+                     "{\"state\":\"%d\", \"door_status\":\"0x%02X\", \"INS_val\":\"%f\", \"occupation_detection_threshold\":\"%lu\", \"stillness_threshold\":\"%lu\", \"timer_status\":\"%lu\", "
                      "\"occupation_detection_timer\":\"%lu\", \"initial_timer\":\"%lu\", \"duration_timer\":\"%lu\", \"stillness_timer\":\"%lu\"}",
-                     state, doorStatus, INSValue, ins_threshold, timer, state0_occupant_detection_timer, state1_max_time, state2_max_duration,
+                     state, doorStatus, INSValue, occupation_detection_ins_threshold, stillness_ins_threshold, timer, state0_occupant_detection_timer, state1_max_time, state2_max_duration,
                      *max_stillness_time);
             Particle.publish("Debug Message", debugMessage, PRIVATE);
             lastDebugPublish = millis();